chore: recent product list -> 공통 productlist로 바꿈
- 어디에서도 사용 가능하도록 바꿈
- 아무 props 안내려주면 디폴트 메인/상품리스트와 동일한 레이아웃을 가져감
- 모바일 작업할 때 모바일 용 간격 적용해 주어야함 (후작업)
