--- src/widgets/recent-product-list/ui/recent-product-list-style.ts
@@ -1,8 +1,0 @@
-import styled from 'styled-components';
-
-export const ProductListContainer = styled.div`
-  display: flex;
-  flex-wrap: wrap;
-  gap: 60px 30px;
-  align-items: center;
-`;
