--- src/widgets/recent-product-list/ui/recent-product-list.tsx
@@ -1,20 +1,0 @@
-import { ProductCardItem } from '@entities/product/product-card-item';
-
-import * as S from './recent-product-list-style';
-
-const RecentProductList = () => {
-  return (
-    <S.ProductListContainer>
-      <ProductCardItem />
-      <ProductCardItem />
-      <ProductCardItem />
-      <ProductCardItem />
-      <ProductCardItem />
-      <ProductCardItem />
-      <ProductCardItem />
-      <ProductCardItem />
-    </S.ProductListContainer>
-  );
-};
-
-export { RecentProductList };
